[ll] core: rework pipeline state description
